# backend/services/api-gateway/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PatientRecords API (stub)",
    "version": "0.1.0",
    "description": "Minimal API spec for development â€” includes auth and patient endpoints."
  },
  "servers": [
    { "url": "http://localhost:5000", "description": "Local dev" }
  ],
...
  "paths": {
...
    "/api/patients/{id}/vitals": {
      "get": {
        "tags": ["Vitals"],
        "summary": "Get vitals for a patient",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "from", "in": "query", "schema": { "type": "string" }, "description": "Start date (inclusive)" },
          { "name": "to", "in": "query", "schema": { "type": "string" }, "description": "End date (inclusive)" },
          { "name": "type", "in": "query", "schema": { "type": "string" }, "description": "Filter by vital type" }
        ],
        "responses": { "200": { "description": "Array of vitals" }, "404": { "description": "Not found" } },
        "security": [ { "bearerAuth": [] } ]
      },
      "post": {
        "tags": ["Vitals"],
        "summary": "Create new vital record",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dateofobservation": { "type": "string" },
                  "observationcode": { "type": "string" },
                  "observationcodesystem": { "type": "string" },
                  "organizationname": { "type": "string" },
                  "vital_description": { "type": "string" },
                  "unit": { "type": "string" },
                  "value": { "type": "string" },
                  "percentile": { "type": "string" }
                },
                "required": ["dateofobservation", "vital_description"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Vital record created" }, "400": { "description": "Bad request" }, "404": { "description": "Patient not found" } },
        "security": [ { "bearerAuth": [] } ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  }
}